int to long

+++ src/Main.java
@@ -1,7 +1,7 @@
 
 public class Main {
     public static void main(String[] args) {
-        int[] sales = new int[2];
+        long[] sales = new long[2];
         sales[0] = 1680;
         sales[1] = 2260;
         SalesManager salesManager = new SalesManager(sales);
